requirements.txt
torch
torchvision
numpy
pillow
matplotlib
scipy
h5py
